Move get_config_file_path to utils.py
