+++ src/ts_backend_check/cli/main.py
@@ -20,36 +20,9 @@ from ts_backend_check.cli.generate_config_file import (
 from ts_backend_check.cli.generate_test_project import generate_test_project
 from ts_backend_check.cli.upgrade import upgrade_cli
 from ts_backend_check.cli.version import get_version_message
+from ts_backend_check.utils import get_config_file_path
 
 # MARK: Base Paths
-
-CWD_PATH = Path.cwd()
-
-
-def get_config_file_path() -> Path:
-    """
-    Get the path to the ts-backend-check configuration file.
-
-    Checks for both .yaml and .yml extensions, preferring .yaml if both exist.
-
-    Returns
-    -------
-    Path
-        The path to the configuration file (.yaml or .yml).
-    """
-    yaml_path = CWD_PATH / ".ts-backend-check.yaml"
-    yml_path = CWD_PATH / ".ts-backend-check.yml"
-
-    # Prefer .yaml if it exists, otherwise check for .yml.
-    if yaml_path.is_file():
-        return yaml_path
-
-    elif yml_path.is_file():
-        return yml_path
-
-    else:
-        # Default to .yaml for new files.
-        return yaml_path
 
 
 def check_files_and_print_results(
