+++ tests/conftest.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 import yaml
 
-from ts_backend_check.cli.main import get_config_file_path
+from ts_backend_check.utils import get_config_file_path
 
 YAML_CONFIG_FILE_PATH = get_config_file_path()
 with open(YAML_CONFIG_FILE_PATH, "r", encoding="utf-8") as file:
